refactor: split github.go into github / comment_cli / review_file
github.go had grown to ~1774 lines mixing three unrelated concerns:
gh-CLI / PR push-pull, the comment-authoring CLI, and review-file I/O.
PR #424's bulk-routing fallback made the entanglement obvious.

Move whole functions, no behavior changes, no signature changes:

- review_file.go: resolveReviewPath, resolveReviewPathFromDaemon,
  pickReviewPath, resolveReviewPathFromSessions, writeCritJSON,
  loadCritJSON, saveCritJSON, clearCritJSON,
  findReviewFileByCommentID, reviewFileContainsComment,
  cjContainsCommentID

- comment_cli.go: appendComment[Scoped], readAnchorFromDisk,
  appendReply, addCommentToCritJSON[Scoped], addReplyToCritJSON,
  BulkCommentEntry (incl. UnmarshalJSON), processBulkEntry and
  helpers, parseLineSpec, bulkAddCommentsToCritJSON[Scoped],
  resolveBulkTarget, addReviewCommentToCritJSON[Scoped],
  addFileCommentToCritJSON[Scoped], appendReviewComment[Scoped],
  appendFileComment[Scoped]

- github.go: keeps only the gh-CLI / PR sync surface
  (detectPR, fetchPRComments, mergeGHComments*, createGHReview,
  push/pull plumbing, truncateStr).

Tests stay in github_test.go for this commit — they all live in
package main and continue to compile and pass against the moved
functions. Splitting the 2700-line test file along the same lines
is mechanical follow-up work and not required for the refactor to
be useful.

go build ./... and go test ./... both pass.
