+++ github.go
@@ -743,109 +743,6 @@ func mergeGHCommentsWithNames(cj *CritJSON, ghComments []ghComment, names userNa
 	return added
 }
 
-// resolveReviewPath returns the full path to the review file for the current context.
-// Resolution order:
-//  1. If outputDir is set, return outputDir/.crit.json (explicit override)
-//  2. Check daemon registry for running sessions matching this cwd
-//  3. If one daemon matches, use its ReviewPath
-//  4. If multiple daemons match, use the one matching current branch
-//  5. If no daemon found, compute the centralized path: ~/.crit/reviews/<key>.json
-func resolveReviewPath(outputDir string) (string, error) {
-	if outputDir != "" {
-		abs, err := filepath.Abs(outputDir)
-		if err != nil {
-			return "", err
-		}
-		return filepath.Join(abs, ".crit.json"), nil
-	}
-
-	cwd, err := resolvedCWD()
-	if err != nil {
-		return "", err
-	}
-
-	if path := resolveReviewPathFromDaemon(cwd); path != "" {
-		return path, nil
-	}
-
-	// No daemon — compute centralized path.
-	branch := ""
-	if vcs := DetectVCS(""); vcs != nil {
-		branch = vcs.CurrentBranch()
-	}
-	key := sessionKey(cwd, branch, nil)
-	path, err := reviewFilePath(key)
-	if err != nil {
-		return "", err
-	}
-
-	return path, nil
-}
-
-// resolveReviewPathFromDaemon checks the daemon registry for a running session
-// and returns its review path. Tries exact CWD match first, then falls back to
-// matching by git repo root (handles subdirectory mismatch — e.g. daemon started
-// from repo/api but crit comment run from repo/).
-func resolveReviewPathFromDaemon(cwd string) string {
-	sessions, _ := listSessionsForCWD(cwd)
-	if path := pickReviewPath(sessions); path != "" {
-		return path
-	}
-
-	// Fallback: match by VCS repo root.
-	if len(sessions) == 0 {
-		vcs := DetectVCS("")
-		if vcs == nil {
-			return ""
-		}
-		if repoRoot, err := vcs.RepoRoot(); err == nil && repoRoot != cwd {
-			repoSessions, _ := listSessionsForRepoRoot(repoRoot)
-			if path := pickReviewPath(repoSessions); path != "" {
-				return path
-			}
-		}
-	}
-	return ""
-}
-
-// pickReviewPath selects a review path from a list of sessions.
-// Returns the path if exactly one session has one, or defers to branch matching for multiple.
-func pickReviewPath(sessions []sessionEntry) string {
-	if len(sessions) == 1 && sessions[0].ReviewPath != "" {
-		return sessions[0].ReviewPath
-	}
-	if len(sessions) > 1 {
-		return resolveReviewPathFromSessions(sessions)
-	}
-	return ""
-}
-
-// resolveReviewPathFromSessions picks the best ReviewPath from multiple daemon sessions.
-// Tries current branch first, then falls back to the first session with a ReviewPath.
-func resolveReviewPathFromSessions(sessions []sessionEntry) string {
-	branch := CurrentBranch()
-	for _, s := range sessions {
-		if s.Branch == branch && s.ReviewPath != "" {
-			return s.ReviewPath
-		}
-	}
-	for _, s := range sessions {
-		if s.ReviewPath != "" {
-			return s.ReviewPath
-		}
-	}
-	return ""
-}
-
-// writeCritJSON resolves the review path and writes a CritJSON via saveCritJSON.
-func writeCritJSON(cj CritJSON, outputDir string) error {
-	path, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-	return saveCritJSON(path, cj)
-}
-
 // ghReplyForPush represents a reply that needs to be posted to GitHub.
 type ghReplyForPush struct {
 	ParentGHID int64
@@ -1069,677 +966,4 @@ func truncateStr(s string, n int) string {
 		return s
 	}
 	return string(r[:n])
-}
-
-// loadCritJSON reads the review file from disk, or returns a fresh CritJSON if the file doesn't exist.
-func loadCritJSON(critPath string) (CritJSON, error) {
-	var cj CritJSON
-	if data, err := os.ReadFile(critPath); err == nil {
-		if err := json.Unmarshal(data, &cj); err != nil {
-			return cj, fmt.Errorf("invalid existing review file: %w", err)
-		}
-	} else if os.IsNotExist(err) {
-		branch := CurrentBranch()
-		cfg := LoadConfig(filepath.Dir(critPath))
-		base := cfg.BaseBranch
-		if base == "" {
-			base = defaultBaseRef()
-		}
-		baseRef, _ := MergeBase(base)
-		cj = CritJSON{
-			Branch:      branch,
-			BaseRef:     baseRef,
-			ReviewRound: 1,
-			Files:       make(map[string]CritJSONFile),
-		}
-	} else {
-		return cj, fmt.Errorf("reading review file: %w", err)
-	}
-	return cj, nil
-}
-
-// saveCritJSON writes the CritJSON struct to disk with pretty-printed JSON
-// and a trailing newline. Uses atomic writes to prevent corruption.
-func saveCritJSON(critPath string, cj CritJSON) error {
-	data, err := json.MarshalIndent(cj, "", "  ")
-	if err != nil {
-		return fmt.Errorf("marshaling review file: %w", err)
-	}
-	// Ensure parent directory exists (centralized path may not exist yet).
-	if err := os.MkdirAll(filepath.Dir(critPath), 0700); err != nil {
-		return fmt.Errorf("creating review directory: %w", err)
-	}
-	return atomicWriteFile(critPath, append(data, '\n'), 0644)
-}
-
-// appendCommentScoped adds a comment to the CritJSON struct in memory with
-// HeadSHA / DiffScope stamping. Does not write to disk.
-// scope.DiffScope == "" produces today's working-tree behavior.
-func appendCommentScoped(cj *CritJSON, filePath string, startLine, endLine int, body, author, userID string, scope inheritedScope) {
-	now := time.Now().UTC().Format(time.RFC3339)
-	cj.UpdatedAt = now
-
-	cf, ok := cj.Files[filePath]
-	if !ok {
-		cf = CritJSONFile{
-			Status:   "modified",
-			Comments: []Comment{},
-		}
-	}
-
-	// Populate anchor from the file on disk.
-	anchor := readAnchorFromDisk(filePath, startLine, endLine)
-
-	c := stampWithFocus(Comment{
-		ID:        randomCommentID(),
-		StartLine: startLine,
-		EndLine:   endLine,
-		Body:      body,
-		Anchor:    anchor,
-		Author:    author,
-		UserID:    userID,
-		CreatedAt: now,
-		UpdatedAt: now,
-	}, scope.asFocus())
-	cf.Comments = append(cf.Comments, c)
-	cj.Files[filePath] = cf
-}
-
-// readAnchorFromDisk reads the file from disk and extracts lines startLine..endLine
-// as the anchor text. Returns empty string on any error or if the file is not found.
-func readAnchorFromDisk(filePath string, startLine, endLine int) string {
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		return ""
-	}
-	return extractAnchor(string(data), startLine, endLine)
-}
-
-// appendReply adds a reply to an existing comment in the CritJSON struct in memory.
-// Returns an error if the comment ID is not found or is ambiguous across files.
-// Searches both file comments and review_comments.
-func appendReply(cj *CritJSON, commentID, body, author, userID string, resolve bool, filterPath string) error {
-	now := time.Now().UTC().Format(time.RFC3339)
-	cj.UpdatedAt = now
-
-	// Check all review comments (not just those starting with "r" — web-fetched ones use "web-N").
-	for i, c := range cj.ReviewComments {
-		if c.ID == commentID {
-			reply := Reply{
-				ID:        randomReplyID(),
-				Body:      body,
-				Author:    author,
-				UserID:    userID,
-				CreatedAt: now,
-			}
-			cj.ReviewComments[i].Replies = append(cj.ReviewComments[i].Replies, reply)
-			cj.ReviewComments[i].UpdatedAt = now
-			if resolve {
-				cj.ReviewComments[i].Resolved = true
-			}
-			return nil
-		}
-	}
-
-	// Search file comments
-	var found bool
-	var foundPaths []string
-	for filePath, cf := range cj.Files {
-		if filterPath != "" && filePath != filterPath {
-			continue
-		}
-		for i, c := range cf.Comments {
-			if c.ID == commentID {
-				foundPaths = append(foundPaths, filePath)
-				if !found {
-					found = true
-					reply := Reply{
-						ID:        randomReplyID(),
-						Body:      body,
-						Author:    author,
-						UserID:    userID,
-						CreatedAt: now,
-					}
-					cf.Comments[i].Replies = append(cf.Comments[i].Replies, reply)
-					cf.Comments[i].UpdatedAt = now
-					if resolve {
-						cf.Comments[i].Resolved = true
-					}
-					cj.Files[filePath] = cf
-				}
-			}
-		}
-	}
-
-	if len(foundPaths) > 1 {
-		return fmt.Errorf("comment %q found in multiple files (%s); use --path <file> to disambiguate",
-			commentID, strings.Join(foundPaths, ", "))
-	}
-	if !found {
-		if filterPath != "" {
-			return fmt.Errorf("comment %q not found in file %q in review file", commentID, filterPath)
-		}
-		return fmt.Errorf("comment %q not found in review file", commentID)
-	}
-	return nil
-}
-
-// addCommentToCritJSONScoped appends a comment to the review file for the given file and line range.
-// Creates the review file if it doesn't exist. Appends to existing comments if it does.
-// Works in both git repos and plain directories (file mode).
-// outputDir overrides the default location (repo root or CWD) when non-empty.
-// scope carries optional HeadSHA / DiffScope stamping; the zero value produces
-// working-tree behavior.
-// userID is documented to support callers that pass an authenticated user; tests
-// pass "" because they don't need that path.
-//
-//nolint:unparam // userID is part of the public contract; tests don't exercise it
-func addCommentToCritJSONScoped(filePath string, startLine, endLine int, body, author, userID, outputDir string, scope inheritedScope) error {
-	critPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-
-	cleaned := filepath.Clean(filePath)
-	if filepath.IsAbs(cleaned) || strings.HasPrefix(cleaned, "..") {
-		return fmt.Errorf("path %q must be relative and within the repository", filePath)
-	}
-
-	cj, err := loadCritJSON(critPath)
-	if err != nil {
-		return err
-	}
-
-	appendCommentScoped(&cj, cleaned, startLine, endLine, body, author, userID, scope)
-	return saveCritJSON(critPath, cj)
-}
-
-// addReplyToCritJSON adds a reply to an existing comment in the review file.
-// It searches all files for the comment ID. If resolve is true, it also marks the comment as resolved.
-func addReplyToCritJSON(commentID, body, author, userID string, resolve bool, outputDir string, filterPath string) error {
-	critPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-
-	cj, err := loadCritJSON(critPath)
-	if err != nil {
-		return err
-	}
-
-	if err := appendReply(&cj, commentID, body, author, userID, resolve, filterPath); err != nil {
-		// Only fall back to scanning when the comment genuinely wasn't found.
-		// Don't fall back for ambiguity errors ("found in multiple files").
-		if strings.Contains(err.Error(), "not found") {
-			if altPath, err2 := findReviewFileByCommentID(commentID, critPath); err2 == nil {
-				altCJ, loadErr := loadCritJSON(altPath)
-				if loadErr != nil {
-					return err // return original error
-				}
-				if replyErr := appendReply(&altCJ, commentID, body, author, userID, resolve, filterPath); replyErr != nil {
-					return err // return original error
-				}
-				fmt.Fprintf(os.Stderr, "Note: comment %s found in %s (not the resolved review file)\n", commentID, filepath.Base(altPath))
-				return saveCritJSON(altPath, altCJ)
-			}
-		}
-		return err
-	}
-	return saveCritJSON(critPath, cj)
-}
-
-// findReviewFileByCommentID scans all review files in ~/.crit/reviews/ for the given
-// comment ID, skipping excludePath. Returns the path if found in exactly one file,
-// or an error wrapping commentID if it's missing or appears in multiple files.
-func findReviewFileByCommentID(commentID string, excludePath string) (string, error) {
-	dir, err := reviewsDir()
-	if err != nil {
-		return "", err
-	}
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return "", fmt.Errorf("comment %q not found in any review file", commentID)
-		}
-		return "", err
-	}
-
-	var matchPath string
-	for _, de := range entries {
-		if !strings.HasSuffix(de.Name(), ".json") {
-			continue
-		}
-		path := filepath.Join(dir, de.Name())
-		if path == excludePath {
-			continue
-		}
-		data, readErr := os.ReadFile(path)
-		if readErr != nil {
-			continue
-		}
-		if reviewFileContainsComment(data, commentID) {
-			if matchPath != "" {
-				return "", fmt.Errorf("comment %q found in multiple review files", commentID)
-			}
-			matchPath = path
-		}
-	}
-	if matchPath == "" {
-		return "", fmt.Errorf("comment %q not found in any review file", commentID)
-	}
-	return matchPath, nil
-}
-
-// reviewFileContainsComment does a quick check if a review JSON file contains
-// a comment with the given ID. Uses string search first as a fast path to
-// avoid parsing files that definitely don't contain the ID.
-func reviewFileContainsComment(data []byte, commentID string) bool {
-	// Fast path: if the ID string doesn't appear at all, skip JSON parsing.
-	if !bytes.Contains(data, []byte(commentID)) {
-		return false
-	}
-	var cj CritJSON
-	if err := json.Unmarshal(data, &cj); err != nil {
-		return false
-	}
-	for _, c := range cj.ReviewComments {
-		if c.ID == commentID {
-			return true
-		}
-	}
-	for _, cf := range cj.Files {
-		for _, c := range cf.Comments {
-			if c.ID == commentID {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// clearCritJSON removes the review file from the resolved path or outputDir.
-func clearCritJSON(outputDir string) error {
-	critPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-	if err := os.Remove(critPath); err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	return nil
-}
-
-// BulkCommentEntry represents one entry in a bulk comment JSON array.
-// Supports review-level, file-level, line-level comments, and replies.
-type BulkCommentEntry struct {
-	// New comment fields
-	File     string `json:"file,omitempty"`
-	Path     string `json:"path,omitempty"`     // alias for File
-	Line     int    `json:"-"`                  // parsed from "line" (int or string like "45-47")
-	LineSpec string `json:"-"`                  // string line spec like "45-47" (from "line" field)
-	EndLine  int    `json:"end_line,omitempty"` // defaults to Line if omitted
-	Body     string `json:"body"`
-	Author   string `json:"author,omitempty"` // overrides per-entry; falls back to global
-	Scope    string `json:"scope,omitempty"`  // "review", "file", or "" (inferred)
-
-	// Reply fields
-	ReplyTo string `json:"reply_to,omitempty"`
-	Resolve bool   `json:"resolve,omitempty"`
-}
-
-// UnmarshalJSON implements custom JSON unmarshaling for BulkCommentEntry
-// to handle the "line" field being either an int (42) or a string ("45-47").
-func (e *BulkCommentEntry) UnmarshalJSON(data []byte) error {
-	// Use an alias to avoid infinite recursion
-	type Alias BulkCommentEntry
-	aux := &struct {
-		Line json.RawMessage `json:"line,omitempty"`
-		*Alias
-	}{
-		Alias: (*Alias)(e),
-	}
-
-	if err := json.Unmarshal(data, aux); err != nil {
-		return err
-	}
-
-	if len(aux.Line) > 0 {
-		// Try int first
-		var lineInt int
-		if err := json.Unmarshal(aux.Line, &lineInt); err == nil {
-			e.Line = lineInt
-			return nil
-		}
-		// Try string
-		var lineStr string
-		if err := json.Unmarshal(aux.Line, &lineStr); err == nil {
-			e.LineSpec = lineStr
-			return nil
-		}
-	}
-	return nil
-}
-
-// processBulkEntry routes a single bulk comment entry to the appropriate
-// authoring helper. globalAuthor is used when an entry doesn't specify its own
-// author. scope is stamped on every authored comment (empty = today's behavior).
-func processBulkEntry(cj *CritJSON, i int, e BulkCommentEntry, globalAuthor, globalUserID string, scope inheritedScope) error {
-	if e.Body == "" {
-		return fmt.Errorf("entry %d: body is required", i)
-	}
-
-	author := e.Author
-	if author == "" {
-		author = globalAuthor
-	}
-	// UserID always comes from the local config — entry-level override would
-	// be a spoof vector. The CLI never trusts a payload-supplied user_id.
-	userID := globalUserID
-
-	if e.ReplyTo != "" {
-		if err := appendReply(cj, e.ReplyTo, e.Body, author, userID, e.Resolve, e.File); err != nil {
-			return fmt.Errorf("entry %d: %w", i, err)
-		}
-		return nil
-	}
-
-	if e.Scope == "review" || (e.File == "" && e.Path == "" && e.Line <= 0 && e.LineSpec == "") {
-		return processBulkReviewEntry(cj, i, e, author, userID, scope)
-	}
-
-	return processBulkFileOrLineEntry(cj, i, e, author, userID, scope)
-}
-
-func processBulkReviewEntry(cj *CritJSON, i int, e BulkCommentEntry, author, userID string, scope inheritedScope) error {
-	if e.Line > 0 || e.LineSpec != "" {
-		return fmt.Errorf("entry %d: file is required for new comments", i)
-	}
-	if e.Scope != "review" && (e.File != "" || e.Path != "") {
-		return fmt.Errorf("entry %d: file is required for new comments", i)
-	}
-	appendReviewCommentScoped(cj, e.Body, author, userID, scope)
-	return nil
-}
-
-func processBulkFileOrLineEntry(cj *CritJSON, i int, e BulkCommentEntry, author, userID string, scope inheritedScope) error {
-	filePath := e.File
-	if filePath == "" {
-		filePath = e.Path
-	}
-	if filePath == "" {
-		return fmt.Errorf("entry %d: file is required for new comments", i)
-	}
-
-	cleaned := filepath.Clean(filePath)
-	if filepath.IsAbs(cleaned) || strings.HasPrefix(cleaned, "..") {
-		return fmt.Errorf("entry %d: path %q must be relative and within the repository", i, filePath)
-	}
-
-	if e.Scope == "file" {
-		appendFileCommentScoped(cj, cleaned, e.Body, author, userID, scope)
-		return nil
-	}
-
-	if e.Line <= 0 && e.LineSpec == "" {
-		if e.Path != "" && e.File == "" {
-			appendFileCommentScoped(cj, cleaned, e.Body, author, userID, scope)
-			return nil
-		}
-		return fmt.Errorf("entry %d: line must be > 0", i)
-	}
-
-	return processBulkLineComment(cj, i, e, cleaned, author, userID, scope)
-}
-
-func processBulkLineComment(cj *CritJSON, i int, e BulkCommentEntry, cleaned, author, userID string, scope inheritedScope) error {
-	startLine := e.Line
-	endLine := e.EndLine
-
-	if e.LineSpec != "" && startLine == 0 {
-		var err error
-		startLine, endLine, err = parseLineSpec(e.LineSpec)
-		if err != nil {
-			return fmt.Errorf("entry %d: invalid line spec %q", i, e.LineSpec)
-		}
-	}
-
-	if startLine <= 0 {
-		return fmt.Errorf("entry %d: line must be > 0", i)
-	}
-	if endLine == 0 {
-		endLine = startLine
-	}
-
-	appendCommentScoped(cj, cleaned, startLine, endLine, e.Body, author, userID, scope)
-	return nil
-}
-
-func parseLineSpec(spec string) (start, end int, err error) {
-	if dashIdx := strings.Index(spec, "-"); dashIdx >= 0 {
-		s, err1 := strconv.Atoi(spec[:dashIdx])
-		e, err2 := strconv.Atoi(spec[dashIdx+1:])
-		if err1 != nil || err2 != nil {
-			if err1 != nil {
-				return 0, 0, err1
-			}
-			return 0, 0, err2
-		}
-		return s, e, nil
-	}
-	n, err := strconv.Atoi(spec)
-	if err != nil {
-		return 0, 0, err
-	}
-	return n, n, nil
-}
-
-// bulkAddCommentsToCritJSONScoped writes a batch of comments to a single review
-// file with scope stamping applied to every entry.
-//
-// Target resolution: a bulk call always writes to a single review file. If any
-// entry uses reply_to and none of the referenced IDs live in the cwd-resolved
-// primary file, the entire bulk is redirected to the alt file that contains
-// them. New comments in the same bulk ride along. If reply IDs split across
-// multiple review files (or some land in primary and others elsewhere), the
-// call is rejected — callers should split into per-file bulks.
-//
-//nolint:unparam // globalUserID is part of the public contract; tests don't exercise it
-func bulkAddCommentsToCritJSONScoped(entries []BulkCommentEntry, globalAuthor, globalUserID string, outputDir string, scope inheritedScope) error {
-	if len(entries) == 0 {
-		return fmt.Errorf("no comment entries provided")
-	}
-
-	primaryPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-
-	primary, err := loadCritJSON(primaryPath)
-	if err != nil {
-		return err
-	}
-
-	targetPath, targetCJ, redirected, err := resolveBulkTarget(entries, primaryPath, primary)
-	if err != nil {
-		return err
-	}
-
-	for i, e := range entries {
-		if err := processBulkEntry(&targetCJ, i, e, globalAuthor, globalUserID, scope); err != nil {
-			return err
-		}
-	}
-
-	if err := saveCritJSON(targetPath, targetCJ); err != nil {
-		return err
-	}
-	if redirected {
-		fmt.Fprintf(os.Stderr, "Note: replies routed to %s (not the cwd-resolved review file)\n", filepath.Base(targetPath))
-	}
-	return nil
-}
-
-// resolveBulkTarget picks the single review file that this bulk call should
-// write to. Returns the path, the loaded CritJSON to mutate, and whether the
-// target differs from the cwd-resolved primary.
-func resolveBulkTarget(entries []BulkCommentEntry, primaryPath string, primary CritJSON) (string, CritJSON, bool, error) {
-	var replyIDs []string
-	seen := map[string]bool{}
-	for _, e := range entries {
-		if e.ReplyTo == "" || seen[e.ReplyTo] {
-			continue
-		}
-		seen[e.ReplyTo] = true
-		replyIDs = append(replyIDs, e.ReplyTo)
-	}
-
-	if len(replyIDs) == 0 {
-		return primaryPath, primary, false, nil
-	}
-
-	var inPrimary, missing []string
-	for _, id := range replyIDs {
-		if cjContainsCommentID(&primary, id) {
-			inPrimary = append(inPrimary, id)
-		} else {
-			missing = append(missing, id)
-		}
-	}
-
-	if len(missing) == 0 {
-		return primaryPath, primary, false, nil
-	}
-	if len(inPrimary) > 0 {
-		return "", CritJSON{}, false, fmt.Errorf(
-			"bulk targets multiple review files: %v exist in %s, but %v do not — split into per-file bulks",
-			inPrimary, filepath.Base(primaryPath), missing,
-		)
-	}
-
-	// None in primary — every reply ID must live in the same alt file.
-	var altPath string
-	for _, id := range missing {
-		path, err := findReviewFileByCommentID(id, primaryPath)
-		if err != nil {
-			return "", CritJSON{}, false, fmt.Errorf("reply target %s: %w", id, err)
-		}
-		if altPath == "" {
-			altPath = path
-			continue
-		}
-		if path != altPath {
-			return "", CritJSON{}, false, fmt.Errorf(
-				"bulk targets multiple review files: %s in %s, %s in %s — split into per-file bulks",
-				missing[0], filepath.Base(altPath), id, filepath.Base(path),
-			)
-		}
-	}
-
-	altCJ, err := loadCritJSON(altPath)
-	if err != nil {
-		return "", CritJSON{}, false, fmt.Errorf("load %s: %w", filepath.Base(altPath), err)
-	}
-	return altPath, altCJ, true, nil
-}
-
-// cjContainsCommentID reports whether the given comment ID exists in the
-// in-memory CritJSON, across review-level and per-file comments.
-func cjContainsCommentID(cj *CritJSON, id string) bool {
-	for _, c := range cj.ReviewComments {
-		if c.ID == id {
-			return true
-		}
-	}
-	for _, cf := range cj.Files {
-		for _, c := range cf.Comments {
-			if c.ID == id {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// addReviewCommentToCritJSONScoped adds a review-level comment to the review
-// file with scope stamping.
-func addReviewCommentToCritJSONScoped(body, author, userID, outputDir string, scope inheritedScope) error {
-	critPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-
-	cj, err := loadCritJSON(critPath)
-	if err != nil {
-		return err
-	}
-
-	appendReviewCommentScoped(&cj, body, author, userID, scope)
-	return saveCritJSON(critPath, cj)
-}
-
-// addFileCommentToCritJSONScoped adds a file-level comment to the review file
-// with scope stamping.
-func addFileCommentToCritJSONScoped(filePath, body, author, userID, outputDir string, scope inheritedScope) error {
-	critPath, err := resolveReviewPath(outputDir)
-	if err != nil {
-		return err
-	}
-
-	cleaned := filepath.Clean(filePath)
-	if filepath.IsAbs(cleaned) || strings.HasPrefix(cleaned, "..") {
-		return fmt.Errorf("path %q must be relative and within the repository", filePath)
-	}
-
-	cj, err := loadCritJSON(critPath)
-	if err != nil {
-		return err
-	}
-
-	appendFileCommentScoped(&cj, cleaned, body, author, userID, scope)
-	return saveCritJSON(critPath, cj)
-}
-
-// appendReviewCommentScoped adds a review-level comment to the CritJSON struct
-// in memory, stamping DiffScope/HeadSHA from scope.
-func appendReviewCommentScoped(cj *CritJSON, body, author, userID string, scope inheritedScope) {
-	now := time.Now().UTC().Format(time.RFC3339)
-	cj.UpdatedAt = now
-
-	c := stampWithFocus(Comment{
-		ID:        randomReviewCommentID(),
-		Body:      body,
-		Author:    author,
-		UserID:    userID,
-		Scope:     "review",
-		CreatedAt: now,
-		UpdatedAt: now,
-	}, scope.asFocus())
-	cj.ReviewComments = append(cj.ReviewComments, c)
-}
-
-// appendFileCommentScoped adds a file-level comment (scope: "file", lines: 0)
-// to the CritJSON struct in memory, stamping DiffScope/HeadSHA from scope.
-func appendFileCommentScoped(cj *CritJSON, filePath, body, author, userID string, scope inheritedScope) {
-	now := time.Now().UTC().Format(time.RFC3339)
-	cj.UpdatedAt = now
-
-	cf, ok := cj.Files[filePath]
-	if !ok {
-		cf = CritJSONFile{
-			Status:   "modified",
-			Comments: []Comment{},
-		}
-	}
-
-	c := stampWithFocus(Comment{
-		ID:        randomCommentID(),
-		Body:      body,
-		Author:    author,
-		UserID:    userID,
-		Scope:     "file",
-		CreatedAt: now,
-		UpdatedAt: now,
-	}, scope.asFocus())
-	cf.Comments = append(cf.Comments, c)
-	cj.Files[filePath] = cf
 }
